Move Autoscaler to file
